requirements.txt
random-text-generator